Addded CSS in three different ways.
Created a styles.css file containing external CSS. To be more specific
I created a class called 'div'. and added this class to the <div> element.

I created internal CSS in the <head> section. In this section I also created
a class and named it 'p'.

Lastly used inline CSS on the <button> element